+++ src/components/TopPlayers/TopPlayers.jsx
@@ -1,100 +1,10 @@
 import React, { useEffect, useState } from 'react'
-import { API_PORT } from '../../redux/constans'
+
 import Player from '../Player/Player'
-import ImageTest from './test.png'
+
 import './TopPlayers.scss'
 
-const TopPlayers = () => {
-
-    const [errorThree, setErrorThree] = useState(false)
-    const [errorTen, setErrorTen] = useState(false)
-    const [topthree, setTopthree] = useState([
-        {
-            position: '1°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '2°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '3°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        }
-    ])
-    ///////////DATA FALSA
-    const [topten, setTopten] = useState([
-        {
-            position: '4°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '5°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '6°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '7°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '8°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '9°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        },
-        {
-            position: '10°',
-            img: ImageTest,
-            name: 'Test',
-            ranking: 43
-        }
-    ])
-    //////////////DATA FALSA
-    //TESTING, NO BORRAR
-    const testingTopThree = () => {
-        fetch(`http://localhost:${API_PORT}/topthree`)
-            .then(data => data.json())
-            .then(res => setTopthree(res))
-            .catch(e => setErrorThree(true))
-    }
-
-    //TESTING, NO BORRAR
-    const testingTopTen = () => {
-        fetch(`http://localhost:${API_PORT}/topten`)
-            .then(data => data.json())
-            .then(res => setTopten(res))
-            .catch(e => setErrorTen(true))
-    }
-
-    useEffect(() => {
-        //TESTING, NO BORRAR
-        testingTopThree()
-        testingTopTen()
-    }, [])
+const TopPlayers = ({topthree, topten}) => {
 
     return (
         <div className='top-players'>
